requirements.txt
fastapi==0.111.0
langchain==0.3.14
langchain_chroma==0.2.0
langchain_community==0.3.14
langchain_core==0.3.30
langchain_groq==0.2.3
langchain_openai==0.3.0
langchain_text_splitters==0.3.5
langsmith==0.2.11
python-dotenv==1.0.1
pymupdf==1.24.9